Modifying Code Alarms
issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I7BJQV

+++ ecmascript/js_hclass.cpp
@@ -695,7 +695,7 @@ bool JSHClass::DumpForProfile(JSHClass *rootHClass, CMap<CString, TrackType> &in
     }
 
     LayoutInfo *layout = LayoutInfo::Cast(rootHClass->GetLayout().GetTaggedObject());
-    int element = rootHClass->NumberOfProps();
+    int element = static_cast<int>(rootHClass->NumberOfProps());
     for (int i = 0; i < element; i++) {
         layout->DumpFieldIndexForProfile(i, infos);
     }
